+++ CHANGELOG.md
@@ -8,6 +8,19 @@ the following form: YYYY.0M.0D.
 
 ## [Unreleased]
 
+## 2025.11.21
+
+### Added
+
+- Designed kernel implementation for RollForSandwich component
+- Includes all kernel and standard methods
+- Representation invariant: bread, protein, cheese, wildCard, and sauce never null
+- Abstraction function where the sandwich is represented by these 5 parts
+
+### Updated
+
+- Changed design to include a kernel implementation with clear, equals, and hashCode methods
+
 ## [2025.9.18]
 - Brainstormed 3 possible components to design
 1. MusicQueue - queue based music playlist